Fixes #12572 - refactor openssl settings
